# docs/openapi-academic-triage.yaml
openapi: 3.0.3
info:
  title: Academic Triage & Request Management System
  description: |
    REST API for the Academic Triage System of the Ingeniería de Sistemas y Computación
    at Universidad del Quindío.

    ## Architecture
    - **Backend**: Spring Boot (Java) with Hexagonal Architecture
    - **Frontend**: Angular (TypeScript)
    - **Authentication**: JWT Bearer Token
    - **AI**: Optional LLM integration for classification suggestions and summaries

    ## System Roles
    | Role | Description |
    |------|-------------|
    | `ADMIN` | Configures catalogs, manages users, defines business rules, and can reject requests |
    | `STAFF` | Classifies, prioritizes, attends, and closes requests |
    | `STUDENT` | Submits requests, can cancel their own requests, and checks status |

    ## Request Lifecycle
    ```
    REGISTERED → CLASSIFIED → IN_PROGRESS → ATTENDED → CLOSED
    REGISTERED → REJECTED
    REGISTERED/CLASSIFIED → CANCELLED
    ```
    Transitions only move forward; terminal states (`CLOSED`, `CANCELLED`, `REJECTED`) are immutable.

    ## Authorization Model
    All endpoints require a valid JWT token unless explicitly marked as public.
    Role-based access is enforced per endpoint as documented.
  version: 1.0.0
  contact:
    name: Development Team
    email: diegoa.floresq@uqvirtual.edu.co
  license:
    name: MIT
servers:
  - url: http://localhost:8080/api/v1
    description: Local development server
...
    description: Production server
...
security:
  - bearerAuth: []
paths:
...
  /users:
    get:
      tags: [Users]
      summary: List users
      description: |
        Returns a paginated list of system users with optional filters.

        **Allowed roles**: ADMIN.
      operationId: listUsers
      parameters:
        - name: role
          in: query
          description: Filter by role
          schema:
            $ref: '#/components/schemas/RoleEnum'
        - name: active
          in: query
          description: Filter by active status
          schema:
            type: boolean
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
      responses:
        '200':
          description: Paginated list of users
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedUserResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /users/{userId}:
    get:
      tags: [Users]
      summary: Get user details
      description: |
        Returns the full profile of a specific user.

        **Allowed roles**: ADMIN, or the user themselves.
      operationId: getUserById
      parameters:
        - $ref: '#/components/parameters/UserIdPath'
      responses:
        '200':
          description: User details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags: [Users]
      summary: Update a user
      description: |
        Updates user information including activation/deactivation and role changes.

        **Allowed roles**: ADMIN.
      operationId: updateUser
      parameters:
        - $ref: '#/components/parameters/UserIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserBody'
      responses:
        '200':
          description: User updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ═══════════════════════════════════════════
  # CATALOGS
  # ═══════════════════════════════════════════
...
  /business-rules:
    get:
      tags: [Business Rules]
      summary: List business rules
      description: |
        Returns all prioritization rules defined in the system.

        **Allowed roles**: ADMIN, STAFF.
      operationId: listBusinessRules
      parameters:
        - name: active
          in: query
          description: Filter by active status
          schema:
            type: boolean
        - name: conditionType
          in: query
          description: Filter by condition type
          schema:
            $ref: '#/components/schemas/ConditionTypeEnum'
      responses:
        '200':
          description: List of business rules
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BusinessRuleResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

    post:
      tags: [Business Rules]
      summary: Create a business rule
      description: |
        Defines a new rule for automatic priority suggestion based on request type,
        impact level, or deadline proximity.

        **Allowed roles**: ADMIN.
      operationId: createBusinessRule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBusinessRuleBody'
            example:
              name: "Urgent seat requests"
              description: "Seat requests with a deadline under 5 days receive HIGH priority."
              conditionType: "REQUEST_TYPE_AND_DEADLINE"
              conditionValue: "5"
              requestTypeId: 4
              resultingPriority: "HIGH"
              active: true
      responses:
        '201':
          description: Business rule created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BusinessRuleResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /business-rules/{ruleId}:
    get:
      tags: [Business Rules]
      summary: Get a business rule by ID
      operationId: getBusinessRuleById
      parameters:
        - $ref: '#/components/parameters/RuleIdPath'
      responses:
        '200':
          description: Business rule details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BusinessRuleResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags: [Business Rules]
      summary: Update a business rule
      description: |
        **Allowed roles**: ADMIN.
      operationId: updateBusinessRule
      parameters:
        - $ref: '#/components/parameters/RuleIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBusinessRuleBody'
      responses:
        '200':
          description: Business rule updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BusinessRuleResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Business Rules]
      summary: Deactivate a business rule
      description: |
        Performs a soft delete by setting the rule to inactive.

        **Allowed roles**: ADMIN.
      operationId: deactivateBusinessRule
      parameters:
        - $ref: '#/components/parameters/RuleIdPath'
      responses:
        '204':
          description: Business rule deactivated
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ═══════════════════════════════════════════
  # AI ASSISTANT
  # ═══════════════════════════════════════════
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login
  parameters:
...
    UserIdPath:
      name: userId
      in: path
      required: true
      description: Unique identifier of the user
      schema:
        type: integer
        format: int64
...
    RuleIdPath:
      name: ruleId
      in: path
      required: true
      description: Unique identifier of the business rule
      schema:
        type: integer
        format: int64
...
    PageParam:
      name: page
      in: query
      description: Page number (zero-based)
      schema:
        type: integer
        default: 0
        minimum: 0
    SizeParam:
      name: size
      in: query
      description: Number of items per page
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
...
  responses:
...
    BadRequest:
      description: Invalid input data
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetailResponse'
          example:
            status: 400
            title: "Bad Request"
            detail: "The request description is required."
    Unauthorized:
      description: Not authenticated or JWT token is invalid/expired
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetailResponse'
          example:
            status: 401
            title: "Unauthorized"
            detail: "Invalid or expired JWT token."
    Forbidden:
      description: Insufficient permissions for this operation
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetailResponse'
          example:
            status: 403
            title: "Forbidden"
            detail: "You do not have permission to perform this operation. Required role: STAFF."
    NotFound:
      description: Resource not found
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetailResponse'
          example:
            status: 404
            title: "Not Found"
            detail: "Request with ID 999 not found."
  schemas:
...
    RoleEnum:
      type: string
      enum: [ADMIN, STAFF, STUDENT]
      description: System roles
...
    PriorityEnum:
      type: string
      enum: [HIGH, MEDIUM, LOW]
      description: Priority levels
    ConditionTypeEnum:
      type: string
      enum: [REQUEST_TYPE, IMPACT_LEVEL, DEADLINE, REQUEST_TYPE_AND_DEADLINE]
      description: Types of conditions evaluated by business rules

    # ── Authentication ──
...
    UserResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        username:
          type: string
          example: "jperez"
        email:
          type: string
          example: "jperez@uniquindio.edu.co"
        firstName:
          type: string
          example: "Juan Carlos"
        lastName:
          type: string
          example: "Pérez López"
        identification:
          type: string
          example: "1094123456"
        role:
          $ref: '#/components/schemas/RoleEnum'
        active:
          type: boolean
          example: true
    UpdateUserBody:
      type: object
      properties:
        firstName:
          type: string
          maxLength: 75
        lastName:
          type: string
          maxLength: 75
        email:
          type: string
          format: email
        role:
          $ref: '#/components/schemas/RoleEnum'
        active:
          type: boolean
    PagedUserResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/UserResponse'
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
        currentPage:
          type: integer
        pageSize:
          type: integer

    # ── Requests ──
...
    RequestTypeResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        name:
          type: string
          example: "Course Registration"
        description:
          type: string
          example: "Request to enroll in courses for the current academic period."
        active:
          type: boolean
...
    BusinessRuleResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
          example: "Urgent seat requests"
        description:
          type: string
          example: "Seat requests with a deadline under 5 days receive HIGH priority."
        conditionType:
          $ref: '#/components/schemas/ConditionTypeEnum'
        conditionValue:
          type: string
          description: |
            Scalar canon by condition type: REQUEST_TYPE = decimal id string of requestType;
            DEADLINE = non-negative days; IMPACT_LEVEL = HIGH|MEDIUM|LOW;
            REQUEST_TYPE_AND_DEADLINE = non-negative days (requestType via requestTypeId).
          example: "4"
        resultingPriority:
          $ref: '#/components/schemas/PriorityEnum'
        requestType:
          $ref: '#/components/schemas/RequestTypeResponse'
          nullable: true
          description: The request type this rule applies to, if any
        active:
          type: boolean
    CreateBusinessRuleBody:
      type: object
      required: [name, conditionType, conditionValue, resultingPriority]
      properties:
        name:
          type: string
          maxLength: 150
        description:
          type: string
          maxLength: 500
        conditionType:
          $ref: '#/components/schemas/ConditionTypeEnum'
        conditionValue:
          type: string
          description: Scalar condition value; semantics depend on conditionType (see BusinessRuleResponse).
        resultingPriority:
          $ref: '#/components/schemas/PriorityEnum'
        requestTypeId:
          type: integer
          format: int64
          nullable: true
          description: Required for REQUEST_TYPE and REQUEST_TYPE_AND_DEADLINE; must be null for DEADLINE and IMPACT_LEVEL
        active:
          type: boolean
          default: true
...
    ProblemDetailResponse:
      type: object
      properties:
        type:
          type: string
          format: uri
          nullable: true
          example: about:blank
        title:
          type: string
          example: Bad Request
        status:
          type: integer
          example: 400
        detail:
          type: string
          example: La solicitud contiene errores de validación
        instance:
          type: string
          format: uri
          nullable: true
        fieldErrors:
          type: array
          nullable: true
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
          description: Per-field validation errors when bean validation fails